@@ -14,7 +14,10 @@
   },
   "dependencies": {
     "babel": "^5.6.14",
+    "babel-polyfill": "^6.2.0",
     "classnames": "^2.1.2",
+    "eslint": "^1.10.1",
+    "eslint-plugin-react": "^3.10.0",
     "flux": "^2.0.3",
     "lazy.js": "^0.4.0",
     "moment": "^2.9.0",
@@ -28,15 +31,20 @@
     "underscore": "^1.7.0"
   },
   "devDependencies": {
-    "babel-core": "^5.5.8",
-    "babel-eslint": "^4.1.1",
-    "babel-loader": "^5.1.4",
+    "babel-core": "^6.2.1",
+    "babel-eslint": "^4.1.6",
+    "babel-loader": "^6.2.0",
+    "babel-preset-es2015": "^6.1.18",
+    "babel-preset-react": "^6.1.18",
+    "babel-preset-stage-0": "^6.1.18",
     "browser-sync": "^2.7.12",
     "chai": "^3.0.0",
     "chalk": "^0.5.1",
     "cheerio": "^0.18.0",
     "envify": "^3.4.0",
-    "eslint-plugin-react": "^2.5.0",
+    "eslint": "^1.10.0",
+    "eslint-loader": "^1.1.1",
+    "eslint-plugin-react": "^3.10.0",
     "gulp": "^3.9.0",
     "gulp-autoprefixer": "^2.3.1",
     "gulp-connect": "^2.2.0",
@@ -51,22 +59,22 @@
     "gulp-zip": "^3.0.2",
     "ionicons": "^2.0.1",
     "json-loader": "^0.5.2",
-    "mocha": "^2.1.0",
+    "mocha": "^2.3.4",
     "mocha-teamcity-reporter": "0.0.4",
-    "npm-shrinkwrap": "^5.4.0",
+    "npm-shrinkwrap": "^200.4.0",
     "oboe": "^2.1.2",
     "sinon": "^1.12.2",
     "source-map-loader": "^0.1.5",
     "source-sans-pro": "^2.0.10",
     "transform-loader": "^0.2.2",
-    "webpack": "^1.9.11"
+    "webpack": "^1.12.9"
   },
   "scripts": {
     "build": "gulp",
     "clean": "rm -rf dist release",
     "dist": "export GULP_ENV=\"production\" && npm run clean && npm run test && npm run build",
     "shrinkwrap": "npm-shrinkwrap --dev",
-    "test": "mocha --ui exports --reporter src/test/reporter/reporterSwitch --compilers jsx:babel-core/register src/test/**/*.test.* src/test/*.test.*",
+    "test": "mocha --ui exports --reporter src/test/reporter/reporterSwitch --require babel-core/register src/test/**/*.test.* src/test/*.test.*",
     "serve": "export GULP_ENV=\"development\" DISABLE_SOURCE_MAP=false && gulp serve",
     "livereload": "export GULP_ENV=\"development\" DISABLE_SOURCE_MAP=false && gulp livereload"
   }
