@@ -2,7 +2,7 @@ var autoprefixer = require("gulp-autoprefixer");
 var connect = require("gulp-connect");
 var header = require('gulp-header');
 var browserSync = require("browser-sync");
-var eslint = require("gulp-eslint");
+var eslintFormatter = require("eslint/lib/formatters/stylish");
 var gulp = require("gulp");
 var gutil = require("gulp-util");
 var less = require("gulp-less");
@@ -35,7 +35,8 @@ var files = {
   mainCssDist: "main",
   index: "index.html",
   sourceSansPro: "SourceSansPro-Regular.otf.woff",
-  sourceSansProBold: "SourceSansPro-Bold.otf.woff"
+  sourceSansProBold: "SourceSansPro-Bold.otf.woff",
+  eslintRc: "./.eslintrc"
 };
 
 var webpackWatch = false;
@@ -45,16 +46,30 @@ if (process.env.GULP_ENV === "development") {
 
 var webpackConfig = {
   entry: dirs.js + "/" + files.mainJs + ".jsx",
+  eslint: {
+    configFile: files.eslintRc,
+    formatter: eslintFormatter
+  },
   output: {
     path: path.resolve(dirs.dist),
     filename: files.mainJsDist + ".js"
   },
   module: {
+    preLoaders: [
+      {
+        test: /\.(js|jsx)$/,
+        loader: "eslint-loader",
+        exclude: /node_modules/
+      }
+    ],
     loaders: [
       {
         test: /\.(js|jsx)$/,
-        loader: "babel-loader?cacheDirectory",
-        exclude: /node_modules/
+        loader: "babel",
+        exclude: /node_modules/,
+        query: {
+          cacheDirectory: true
+        }
       },
       {
         test: /\.json$/,
@@ -82,13 +97,16 @@ gulp.task("webpack", function (callback) {
     !process.env.DISABLE_SOURCE_MAP ||
     process.env.DISABLE_SOURCE_MAP === "false") {
     webpackConfig.devtool = "source-map";
-    webpackConfig.module.preLoaders = [
+    if (webpackConfig.module.preLoaders == null) {
+      webpackConfig.module.preLoaders = [];
+    }
+    webpackConfig.module.preLoaders.push(
       {
         test: /\.js$/,
         loader: "source-map-loader",
         exclude: /node_modules/
       }
-    ];
+    );
   }
   // run webpack
   webpack(webpackConfig, function (err, stats) {
@@ -104,25 +122,16 @@ gulp.task("webpack", function (callback) {
       timing: true
     }));
 
-
     if (isFirstRun) {
       // This runs on initial gulp webpack load.
       isFirstRun = false;
       callback();
     } else {
       // This runs after webpack's internal watch rebuild.
-      eslintFn();
       browserSync.reload();
     }
   });
 });
-
-function eslintFn() {
-  return gulp.src([dirs.js + "/**/*.?(js|jsx)"])
-    .pipe(eslint())
-    .pipe(eslint.formatEach("stylish", process.stderr));
-}
-gulp.task("eslint", eslintFn);
 
 gulp.task("less", function () {
   return gulp.src(dirs.styles + "/" + files.mainLess + ".less")
@@ -193,7 +202,7 @@ gulp.task("watch", function () {
   gulp.watch(dirs.fonts + "/**/*.*", ["fonts"]);
 });
 
-gulp.task("replace-js-strings", ["webpack", "eslint", "minify-js"], function () {
+gulp.task("replace-js-strings", ["webpack", "minify-js"], function () {
   return gulp.src(dirs.dist + "/main.js")
     .pipe(replace("@@ENV", process.env.GULP_ENV))
     .pipe(gulp.dest(dirs.dist));
@@ -203,7 +212,6 @@ gulp.task("serve", ["default", "connect:server", "watch"]);
 gulp.task("livereload", ["default", "browsersync", "watch"]);
 
 var tasks = [
-  "eslint",
   "webpack",
   "less",
   "images",
